pyproject.toml
[project]
name = "wallpipe"
version = "0.0.0"
description = "Wallpaper download/curate/filter pipeline"
requires-python = ">=3.11"
dependencies = [
  "pillow>=10.0.0",
  "torch==2.3.1",
  "transformers>=4.44.0",
  "simple-aesthetics-predictor>=0.0.7",
  "gallery-dl>=1.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "pyright>=1.1.385",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "pyright>=1.1.385",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
filterwarnings = [
    "ignore:.*cudaGetDeviceCount.*:UserWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["."]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["common"]

[tool.pyright]
include = ["."]
exclude = ["__pycache__", ".git", ".venv"]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "warning"
reportMissingModuleSource = "warning"
venvPath = "/home/matthias/.pyenv/versions"
venv = "wallpipe"

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "*/site-packages/*",
    "__pycache__/*",
    "*/_remote_module_non_scriptable.py",
]
